+++ lix/packages/client/src/openRepository.ts
@@ -11,8 +11,15 @@ import { createSignal, createEffect } from "./solid.js"
 import type { OptStatus } from "./git/status-list.js"
 import { commit as lixCommit } from "./git/commit.js"
 import { statusList as lixStatusList } from "./git/status-list.js"
+import { checkout as lixCheckout } from "./git/checkout.js"
+
 import isoGit from "../vendored/isomorphic-git/index.js"
+
+// @ts-ignore
 import { modeToFileType } from "./git/helpers.js"
+
+// const checkout = isoGit.checkout
+const checkout = lixCheckout
 
 // TODO: --filter=tree:0 for commit history?
 
@@ -226,21 +233,19 @@ export async function openRepository(
 			await rawFs.rm(placeholder)
 		}
 
-		const res = await isoGit
-			.checkout({
-				fs: withProxy({
-					nodeishFs: rawFs,
-					verbose: debug,
-					description: debug ? "checkout: " + JSON.stringify(thisBatch) : "checkout",
-				}),
-				dir,
-				cache,
-				ref: args.branch,
-				filepaths: thisBatch,
-			})
-			.catch((error) => {
-				console.error({ error, thisBatch })
-			})
+		const res = await checkout({
+			fs: withProxy({
+				nodeishFs: rawFs,
+				verbose: debug,
+				description: debug ? "checkout: " + JSON.stringify(thisBatch) : "checkout",
+			}),
+			dir,
+			cache,
+			ref: args.branch,
+			filepaths: thisBatch,
+		}).catch((error) => {
+			console.error({ error, thisBatch })
+		})
 
 		for (const entry of thisBatch) {
 			checkedOut.add(entry)
@@ -261,7 +266,7 @@ export async function openRepository(
 		if (!experimentalFeatures.lazyClone) {
 			return
 		}
-		await isoGit.checkout({
+		await checkout({
 			fs: withProxy({
 				nodeishFs: rawFs,
 				verbose: debug,
@@ -327,7 +332,7 @@ export async function openRepository(
 		})
 
 		if (gitignoreFiles.length) {
-			await isoGit.checkout({
+			await checkout({
 				fs: withProxy({
 					nodeishFs: rawFs,
 					verbose: debug,
@@ -637,7 +642,7 @@ export async function openRepository(
 				)
 			}
 
-			await isoGit.checkout({
+			await checkout({
 				fs: withProxy({
 					nodeishFs: rawFs,
 					verbose: debug,
@@ -919,7 +924,7 @@ export async function openRepository(
 				experimentalFeatures.lazyClone = false
 
 				true && console.info('checking out "HEAD" after pull')
-				await isoGit.checkout({
+				await checkout({
 					fs: rawFs,
 					cache,
 					dir,
@@ -927,7 +932,7 @@ export async function openRepository(
 					noCheckout: false,
 				})
 			} else {
-				await isoGit.checkout({
+				await checkout({
 					fs: rawFs,
 					cache,
 					dir,