+++ inlang/source-code/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inlang/cli
 
+## 2.13.4
+
+### Patch Changes
+
+- Updated dependencies [8a8aec7]
+  - @lix-js/client@1.2.0
+  - @inlang/sdk@0.28.3
+
 ## 2.13.3
 
 ### Patch Changes