+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inlang/paraglide-js-adapter-next-example-pages
 
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [1cc9cbc]
+- Updated dependencies [5401f95]
+  - @inlang/paraglide-js@1.3.7
+
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [526b0ba]
+- Updated dependencies [26d2ba1]
+  - @inlang/paraglide-js@1.3.6
+
 ## 0.2.10
 
 ### Patch Changes